refactor: shorten filenames and consolidate trivially small files
