+++ core/lib/docker/args.ts
@@ -1,3 +1,17 @@
+export interface BuildDockerCpArgsOptions {
+  containerName: string;
+  containerPath: string;
+  hostPath: string;
+}
+
+export function buildDockerCpArgs({
+  containerName,
+  containerPath,
+  hostPath,
+}: BuildDockerCpArgsOptions): string[] {
+  return ["cp", `${containerName}:${containerPath}`, hostPath];
+}
+
 /**
  * Build the `docker compose ... up`/`down` argv, shared by setup and run's
  * main/post steps.
