+++ core/lib/errors.ts
@@ -14,3 +14,11 @@ export class ActionError<Code extends string = string> extends Error {
     this.code = code;
   }
 }
+
+/**
+ * Safely extract a message from a caught value of unknown shape — a plain
+ * `Error` most of the time, but `catch` doesn't guarantee that.
+ */
+export function errorMessage(e: unknown): string {
+  return e instanceof Error ? e.message : String(e);
+}
