+++ core/lib/report/outcome/emit.test.ts
@@ -1,8 +1,8 @@
 import { describe, it, beforeEach, afterEach, vi } from "vitest";
 import assert from "node:assert/strict";
 
-import { emitBlockedOutcome } from "./emit-blocked-outcome.ts";
-import { annotateKnownBlocked } from "../build/annotate-known-blocked.ts";
+import { emitBlockedOutcome } from "./emit.ts";
+import { annotateKnownBlocked } from "../build/aggregate.ts";
 import type { ReportDataCommon, GenReportParameters } from "../types.ts";
 
 let prevExitCode: number | string | null | undefined;
