+++ core/lib/report/outcome/emit.ts
@@ -1,7 +1,7 @@
 /** Emits an annotation and sets the exit code for a blocked-connection outcome. */
 import { createAnnotation } from "../../actions/annotation.ts";
 import { describeBlockedOutcome } from "./blocked-outcome.ts";
-import { applyOutcomeAnnotation } from "./apply-outcome-annotation.ts";
+import { applyOutcomeAnnotation } from "./annotate.ts";
 import type { ReportDataCommon } from "../types.ts";
 
 export interface EmitBlockedOutcomeOptions {
