+++ setup/src/post.ts
@@ -3,7 +3,7 @@ import { join, dirname } from "node:path";
 import { fileURLToPath } from "node:url";
 import * as core from "@actions/core";
 import { resolveProjectName } from "../../core/lib/docker/compose-project-name.ts";
-import { buildComposeDownArgs } from "../../core/lib/docker/compose-args.ts";
+import { buildComposeDownArgs } from "../../core/lib/docker/args.ts";
 
 const __dirname = dirname(fileURLToPath(import.meta.url));
